Improved station freight customer view. Fixed logic in removing modules from layout.

+++ SourceCode/Services/Implementations/LayoutService.cs
@@ -28,7 +28,7 @@ namespace ModulesRegistry.Services.Implementations
         /// <param name="layoutId"></param>
         /// <param name="package"></param>
         /// <returns></returns>
-        public async Task<(int Count, string Message)> AddPackageModulesAsync(ClaimsPrincipal? principal, int participantId, int layoutId,  ModulePackage package)
+        public async Task<(int Count, string Message)> AddPackageModulesAsync(ClaimsPrincipal? principal, int participantId, int layoutId, ModulePackage package)
         {
             if (principal.IsAuthenticated())
             {
@@ -41,9 +41,9 @@ namespace ModulesRegistry.Services.Implementations
                     var existing = await dbContext.LayoutModules.SingleOrDefaultAsync(lm => lm.ModuleId == module.Id);
                     if (existing is null)
                     {
-                        var addedModule = new LayoutModule { LayoutId = layoutId, ModuleId = module.Id, ParticipantId = participant.Id, RegisteredTime = TimeProvider.Now };
-                        dbContext.LayoutModules.Add(addedModule);
-                        await AddLayoutStationAsync(dbContext, participantId, layoutId, module);
+                        var layoutModule = new LayoutModule { LayoutId = layoutId, ModuleId = module.Id, ParticipantId = participant.Id, RegisteredTime = TimeProvider.Now };
+                        dbContext.LayoutModules.Add(layoutModule);
+                        await AddLayoutStationAsync(dbContext, participantId, layoutId, module, layoutModule);
                         await dbContext.SaveChangesAsync();
                     }
                     else
@@ -53,7 +53,7 @@ namespace ModulesRegistry.Services.Implementations
                 }
                 return result == 0 ? (-1, Resources.Strings.NoModification) : (1, Resources.Strings.Saved);
 
-                static async Task AddLayoutStationAsync(ModulesDbContext dbContext, int participantId, int layoutId, Module module)
+                static async Task AddLayoutStationAsync(ModulesDbContext dbContext, int participantId, int layoutId, Module module, LayoutModule layoutModule)
                 {
                     if (module.StationId.HasValue)
                     {
@@ -62,6 +62,7 @@ namespace ModulesRegistry.Services.Implementations
                         {
                             var addedStation = new LayoutStation { LayoutId = layoutId, StationId = module.StationId.Value };
                             dbContext.LayoutStations.Add(addedStation);
+                            layoutModule.LayoutStationId = addedStation.Id;
                         }
                     }
                 }
@@ -75,21 +76,35 @@ namespace ModulesRegistry.Services.Implementations
             if (principal.IsAuthenticated())
             {
                 using var dbContext = Factory.CreateDbContext();
-                return await dbContext.LayoutModules.Include(lm => lm.Module).Where(lm => lm.ParticipantId == participantId && lm.LayoutId == layoutId).ToListAsync();
+                return await dbContext.LayoutModules.Include(lm => lm.LayoutStation).Include(lm => lm.Module).Where(lm => lm.ParticipantId == participantId && lm.LayoutId == layoutId).ToListAsync();
             }
             return Array.Empty<LayoutModule>();
         }
 
-        public async Task<(int Count, string Message)> RemoveModuleAsync(ClaimsPrincipal? principal, int layoutModuleId )
+        public async Task<(int Count, string Message)> RemoveModuleAsync(ClaimsPrincipal? principal, int layoutModuleId)
         {
             if (principal.IsAuthenticated())
             {
                 using var dbContext = Factory.CreateDbContext();
-                var existing = await dbContext.LayoutModules.FindAsync(layoutModuleId);
+                var existing = await dbContext.LayoutModules
+                    .Include(lm => lm.LayoutStation).ThenInclude(ls=> ls.LayoutModules)
+                    .SingleOrDefaultAsync( lm => lm.Id == layoutModuleId);
+
                 if (existing is null) return (-1, Resources.Strings.NoModification);
-                dbContext.LayoutModules.Remove(existing);
-                var result = await dbContext.SaveChangesAsync();
-                return result.DeleteResult();
+
+                var stationIsReferredFromOtherLayoutModule = existing.LayoutStation is not null && !existing.LayoutStation.LayoutModules.Any(lm => lm.Id != existing.Id);
+                var layoutStationIsNotUsedInLayout = await dbContext.LayoutStations.Where(ls => ls.Id == existing.LayoutStationId).AnyAsync(ls => !ls.StartingLines.Any() && !ls.EndingLines.Any());
+                if (layoutStationIsNotUsedInLayout)
+                {
+                    if (! stationIsReferredFromOtherLayoutModule)
+                    {
+                        dbContext.LayoutStations.Remove(existing.LayoutStation);
+                    }
+                    dbContext.LayoutModules.Remove(existing);
+                    var result = await dbContext.SaveChangesAsync();
+                    return result.DeleteResult();
+                }
+                Data.Resources.Strings.NotAuthorised.DeleteResult();
             }
             return principal.DeleteNotAuthorized<LayoutModule>();
         }
